Move local `common` protobuf files to `frequenz/api/weather`
The repo-config documentation parts appear to be looking for docs for
`frequenz/api/common` in the `submodules/frequenz-api-common`
directory, which we have temporarily removed.

This change moves the files out of `frequenz/api/common`, in order to
get the documentation working.
