+++ .github/workflows/ci.yaml
@@ -193,8 +193,9 @@ jobs:
 
       - name: Install build dependencies
         run: |
-          python -m pip install -U pip
+          python -m pip install -U pip build
           python -m pip install .[dev-mkdocs]
+          python -m build
           pip freeze
 
       - name: Generate the documentation
