+++ proto/frequenz/api/weather/location.proto
