+++ proto/frequenz/api/weather/pagination/pagination_info.proto
