+++ proto/frequenz/api/weather/pagination/pagination_params.proto
