+++ tests/test_types.py
@@ -13,8 +13,8 @@ from datetime import timedelta
 
 import numpy as np
 from _pytest.logging import LogCaptureFixture
-from frequenz.api.common.location_pb2 import Location as LocationProto
 from frequenz.api.weather import weather_pb2
+from frequenz.api.weather.location_pb2 import Location as LocationProto
 from frequenz.client.weather._types import ForecastFeature, Forecasts, Location
 from google.protobuf.timestamp_pb2 import Timestamp
 from pytest import fixture
